requirement.txt
numpy
torch
torchvision
nibabel
Pillow
scikit-image
pyyaml
lpips
types-PyYAML
